Split out tests
- needed to fix how they were done, as they were dependent on order
